change the type of some extra data from gdynia fields from float to int

+++ app/clients/gdynia_extra_data_client.rb
@@ -14,9 +14,9 @@ class GdyniaExtraDataClient
       # We are interested in the following fields: "w51", "wx2", "wx3", "w68", "w88"
       {
         'rspo' => raw_record.fetch('rspo').to_i,
-        'w51' => raw_record.fetch('w51').to_f,
-        'wx2' => raw_record.fetch('wx2').to_f,
-        'wx3' => raw_record.fetch('wx3').to_f,
+        'w51' => raw_record.fetch('w51').to_i,
+        'wx2' => raw_record.fetch('wx2').to_i,
+        'wx3' => raw_record.fetch('wx3').to_i,
         'w68' => raw_record.fetch('w68').to_f,
         'w88' => raw_record.fetch('w88').to_f
       }
